Database.MongoDB.Connection

Description

Connect to a single server or a replica set of servers

Util

type Secs = Double

Connection

type Pipe = Pipeline

Thread-safe TCP connection with pipelined requests

close :: Pipeline -> IO ()

Close pipe and underlying connection

isClosed :: Pipeline -> IO Bool

Server

data Host

Constructors

Host HostName PortID

data PortID :: *

Constructors

Service String

PortNumber PortNumber

UnixSocket String

defaultPort :: PortID

Default MongoDB port = 27017

host :: HostName -> Host

Host on defaultPort

showHostPort :: Host -> String

Display host as "host:port" TODO: Distinguish Service and UnixSocket port

readHostPort :: String -> Host

Read string "hostname:port" as Host hostname (PortNumber port) or "hostname" as host hostname (default port). Error if string does not match either syntax.

readHostPortM :: Monad m => String -> m Host

Read string "hostname:port" as Host hosthame (PortNumber port) or "hostname" as host hostname (default port). Fail if string does not match either syntax. TODO: handle Service and UnixSocket port

globalConnectTimeout :: IORef Secs

connect (and openReplicaSet) fails if it can't connect within this many seconds (default is 6 seconds). Use 'connect\'' (and 'openReplicaSet\'') if you want to ignore this global and specify your own timeout. Note, this timeout only applies to initial connection establishment, not when reading/writing to the connection.

connect :: Host -> IO Pipe

Connect to Host returning pipelined TCP connection. Throw IOError if connection refused or no response within globalConnectTimeout.

connect' :: Secs -> Host -> IO Pipe

Connect to Host returning pipelined TCP connection. Throw IOError if connection refused or no response within given number of seconds.

Replica Set

type ReplicaSetName = Text

openReplicaSet :: (ReplicaSetName, [Host]) -> IO ReplicaSet

Open connections (on demand) to servers in replica set. Supplied hosts is seed list. At least one of them must be a live member of the named replica set, otherwise fail. The value of globalConnectTimeout at the time of this call is the timeout used for future member connect attempts. To use your own value call 'openReplicaSet\'' instead.

openReplicaSet' :: Secs -> (ReplicaSetName, [Host]) -> IO ReplicaSet

Open connections (on demand) to servers in replica set. Supplied hosts is seed list. At least one of them must be a live member of the named replica set, otherwise fail. Supplied seconds timeout is used for connect attempts to members.

data ReplicaSet

Maintains a connection (created on demand) to each server in the named replica set

primary :: ReplicaSet -> IO Pipe

Return connection to current primary of replica set. Fail if no primary available.

secondaryOk :: ReplicaSet -> IO Pipe

Return connection to a random secondary, or primary if no secondaries available.

routedHost :: ((Host, Bool) -> (Host, Bool) -> IO Ordering) -> ReplicaSet -> IO Pipe

Return a connection to a host using a user-supplied sorting function, which sorts based on a tuple containing the host and a boolean indicating whether the host is primary.

closeReplicaSet :: ReplicaSet -> IO ()

Close all connections to replica set

replSetName :: ReplicaSet -> Text

name of connected replica set