Database.MongoDB.Transport.Tls

Description

This module is for connecting to TLS enabled mongodb servers. ATTENTION!!! Be aware that this module is highly experimental and is barely tested. The current implementation doesn't verify server's identity. It only allows you to connect to a mongodb server using TLS protocol.

Documentation

connect :: HostName -> PortID -> IO Pipe

Connect to mongodb using TLS